requirements.txt
web3
pyyaml==5.4.1
